feat: node execution strategies
