+++ server/services/test-flow/test-flow-executor/sequence-executor.ts
@@ -8,13 +8,10 @@ import { TestFlowReporter } from "../test-flow-reporter.service";
 import { TestFlowService } from "../test-flow.service";
 import {
   AssertionResponse,
-  AssertionResult,
   NodeResponseValidator,
 } from "./node-response-validator";
-import {
-  ITestFlowExecutor,
-  TestFlowExecutor,
-} from "./test-flow-executor.service";
+import { ITestFlowExecutor } from "./test-flow-executor.service";
+import { INodeExecutor } from "../flow-node-executors/executors.types";
 
 const logger = useLog({ dirname: __dirname, filename: __filename });
 
@@ -40,12 +37,8 @@ export type ExecutionContext = Record<string, NodeRunResult>;
 
 export class SequenceExecutor implements ITestFlowExecutor {
   constructor(
-    private readonly requestService: RequestService,
-    private readonly nodeResponseValidator: NodeResponseValidator,
-    private readonly testFlowReporter: TestFlowReporter,
     private readonly testFlowService: TestFlowService,
-    private readonly requestTransformer: RequestTransformer,
-    private readonly testFlowExecutorService: TestFlowExecutor,
+    private readonly nodeExecutionStrategies: Record<string, INodeExecutor>,
   ) {}
 
   async execute(
@@ -69,102 +62,21 @@ export class SequenceExecutor implements ITestFlowExecutor {
     try {
       for (let i = 0; i < sortedNodes.length; i++) {
         const nodeRun = sortedNodes[i];
-        const { subdomain } = nodeRun;
 
-        if (subdomain == null) {
-          throw new Error("subdomain is empty for http node");
-        }
+        const res = await this.nodeExecutionStrategies[nodeRun.type].execute(
+          ownerId,
+          flowId,
+          flowRunId,
+          nodes,
+          nodeRuns,
+          edges,
+          nodeRun,
+          result,
+        );
 
-        if (nodeRun.type === "http") {
-          const {
-            method,
-            url,
-            headers: reqHeaders,
-            body: reqBody,
-          } = this.requestTransformer.transformRequest(nodeRun, result.context);
-
-          const response = await this.requestService.execute({
-            method,
-            url: url ?? "/",
-            headers: reqHeaders || {},
-            body: reqBody,
-            subdomain,
-          });
-
-          const { data: body, headers, status } = response;
-          const assertionResponse = { body, headers, status };
-
-          const assertionResult = await this.nodeResponseValidator.assert(
-            nodeRun,
-            assertionResponse,
-            result.context,
-          );
-
-          await this.testFlowReporter.reportNodeRun(
-            ownerId,
-            flowId,
-            flowRunId,
-            nodeRun,
-            assertionResult,
-            {
-              headers: response?.headers,
-              body:
-                typeof response?.data === "string"
-                  ? response?.data
-                  : JSON.stringify(response?.data ?? "", null, 2),
-              status: response?.status,
-            },
-          );
-
-          const resultItem = {
-            node: nodeRun,
-            response: assertionResponse,
-            assertionResult,
-            success: assertionResult.success,
-          };
-
-          result.context[nodeRun.nodeId] = resultItem;
-
-          if (!assertionResult.success) {
-            result.success = false;
-            break;
-          }
-        } else if (nodeRun.type === "subflow") {
-          if (nodeRun.subFlowId == null) {
-            throw new Error("subflowId is empty for subflow node");
-          }
-          if (nodeRun.subFlowId === flowId) {
-            throw new Error("Flow cannot be a subflow of itself");
-          }
-          const execRes = await this.testFlowExecutorService.execute(
-            ownerId,
-            nodeRun.subFlowId,
-          );
-
-          const assertionResult = await this.nodeResponseValidator.assert(
-            nodeRun,
-            execRes,
-            result.context,
-          );
-
-          await this.testFlowReporter.reportNodeRun(
-            ownerId,
-            flowId,
-            flowRunId,
-            nodeRun,
-            assertionResult,
-            {
-              context: execRes.context,
-              success: assertionResult.success,
-            },
-          );
-
-          result.context[nodeRun.nodeId] = execRes;
-
-          if (!assertionResult.success) {
-            result.success = false;
-            break;
-          }
+        if (!res.success) {
+          result.success = false;
+          break;
         }
       }
